+++ mne_bids/report.py
@@ -229,6 +229,7 @@ def _summarize_participants_tsv(root):
     p_ages = participants_tsv.get('age')
     min_age, max_age = 'n/a', 'n/a'
     mean_age, std_age = 'n/a', 'n/a'
+    print(p_ages,participants_tsv)
     n_age_unknown = len(p_ages) if p_ages else len(p_ids)
     if p_ages:
         # only summarize age if they are numerics